refactor: move pages to their own folder
