+++ src/components/AppDetails/AppDetails.module.css
