+++ src/pages/ManualInstall/ManualInstall.module.css
